lsmeans patch to avoid summary_emm object (#60)
# Pull Request
This PR patches `lsmeans.tern_gee_logistic` so that it creates a true
`data.frame`. The function `emmeans::as.data.frame()` actually creates a
`summary_emm` object, which is an extension of `data.frame` with
annotations added. A new version of **emmeans** added an `rbind()`
method for `summary_emm`, with the result that the `rbind()` call in
line 58 of `lsmeans.R` got dispatched to `emmeans::rbind.summary_emm()`,
initially causing an error until I fixed it to work around it.

In this PR, `confint(prop_emm)` creates a `summary_emm` object, so that
`data.frame()` turns it into a true `data.frame`. Then the `rbind()`
call later on does not get dispatched to **emmeans**, and thus does not
require my special work-around code.